# 🚀  improve performance by not prefixing attributes on annotations