--- a/packages/@atjson/document/test/annotation-test.ts
+++ b/packages/@atjson/document/test/annotation-test.ts
@@ -18,8 +18,8 @@ describe("Annotation", () => {
             start: 0,
             end: 1,
             attributes: {
-              "-test-url": "http://www.example.com/test.jpg",
-              "-test-caption": {
+              url: "http://www.example.com/test.jpg",
+              caption: {
                 content: "An example caption",
                 annotations: [
                   {
@@ -44,8 +44,8 @@ describe("Annotation", () => {
             start: 0,
             end: 1,
             attributes: {
-              "-test-url": "http://www.example.com/test.jpg",
-              "-test-caption": {
+              url: "http://www.example.com/test.jpg",
+              caption: {
                 content: "An example caption",
                 annotations: [
                   {
@@ -70,8 +70,8 @@ describe("Annotation", () => {
             start: 0,
             end: 1,
             attributes: {
-              "-test-url": "http://www.example.com/test.jpg",
-              "-test-caption": {
+              url: "http://www.example.com/test.jpg",
+              caption: {
                 content: "An example caption",
                 annotations: [
                   {