--- a/packages/@atjson/renderer-commonmark/performance/index.js
+++ b/packages/@atjson/renderer-commonmark/performance/index.js
@@ -137,7 +137,8 @@ performance.clearMeasures();
 let fixtures = [
   "alexander-mcqueen.md",
   "lambda-literary-awards.md",
-  "inoa-listings.md" // ~1s
+  "inoa-listings.md", // ~1s
+  "long-list-of-links.md"
 ].map(filename =>
   readFileSync(join(__dirname, "fixtures", filename)).toString()
 );