--- a/packages/@atjson/renderer-graphviz/example.dot
+++ b/packages/@atjson/renderer-graphviz/example.dot
@@ -1,7 +1,7 @@
 digraph atjson{
   node [shape=record];
   root1 [label="{root|{}}" style=filled fillcolor="#222222" fontcolor="#FFFFFF"];
-  link2 [label="{link|{\"url\":\"https://newyorker.com\"}}" style=filled fillcolor="#888888" fontcolor="#FFFFFF"];
+  link2 [label="{link|{\"-test-url\":\"https://newyorker.com\"}}" style=filled fillcolor="#888888" fontcolor="#FFFFFF"];
   text3 [label="{text|The }" style=filled fillcolor="#FFFFFF" fontcolor="#000000"];
   italic4 [label="{italic|{}}" style=filled fillcolor="#888888" fontcolor="#FFFFFF"];
   text5 [label="{text|best}" style=filled fillcolor="#FFFFFF" fontcolor="#000000"];