--- a/packages/@atjson/source-gdocs-paste/src/converter.ts
+++ b/packages/@atjson/source-gdocs-paste/src/converter.ts
@@ -29,13 +29,13 @@ GDocsSource.defineConverterTo(OffsetSource, doc => {
   doc.where({ type: "-gdocs-ts_un" }).set({ type: "-offset-underline" });
   doc.where({ type: "-gdocs-ts_st" }).set({ type: "-offset-strikethrough" });
   doc
-    .where({ type: "-gdocs-ts_va", attributes: { "-gdocs-va": "sub" } })
+    .where({ type: "-gdocs-ts_va", attributes: { va: "sub" } })
     .set({ type: "-offset-subscript" })
-    .unset("attributes.-gdocs-va");
+    .unset("attributes.va");
   doc
-    .where({ type: "-gdocs-ts_va", attributes: { "-gdocs-va": "sup" } })
+    .where({ type: "-gdocs-ts_va", attributes: { va: "sup" } })
     .set({ type: "-offset-superscript" })
-    .unset("attributes.-gdocs-va");
+    .unset("attributes.va");
 
   doc
     .where({ type: "-gdocs-horizontal_rule" })
@@ -52,21 +52,21 @@ GDocsSource.defineConverterTo(OffsetSource, doc => {
   doc
     .where({ type: "-gdocs-ps_hd" })
     .set({ type: "-offset-heading" })
-    .rename({ attributes: { "-gdocs-level": "-offset-level" } });
+    .rename({ attributes: { level: "level" } });
 
   // b_gt: 9 indicates an unordered list, but ordered lists have a variety of b_gt values
   doc
-    .where({ type: "-gdocs-list", attributes: { "-gdocs-ls_b_gt": 9 } })
-    .set({ type: "-offset-list", attributes: { "-offset-type": "bulleted" } });
+    .where({ type: "-gdocs-list", attributes: { ls_b_gt: 9 } })
+    .set({ type: "-offset-list", attributes: { type: "bulleted" } });
   doc
     .where({ type: "-gdocs-list" })
-    .set({ type: "-offset-list", attributes: { "-offset-type": "numbered" } });
+    .set({ type: "-offset-list", attributes: { type: "numbered" } });
   doc.where({ type: "-gdocs-list_item" }).set({ type: "-offset-list-item" });
 
   doc
     .where({ type: "-gdocs-lnks_link" })
     .set({ type: "-offset-link" })
-    .rename({ attributes: { "-gdocs-ulnk_url": "-offset-url" } });
+    .rename({ attributes: { ulnk_url: "url" } });
 
   doc
     .where({ type: "-offset-list" })