--- a/packages/@atjson/source-gdocs-paste/src/link-styles.ts
+++ b/packages/@atjson/source-gdocs-paste/src/link-styles.ts
@@ -29,8 +29,8 @@ export default function extractLinkStyles(
         start: i,
         end: -1,
         attributes: {
-          "-gdocs-ulnk_url": link.lnks_link.ulnk_url,
-          "-gdocs-lnk_type": link.lnks_link.lnk_type
+          ulnk_url: link.lnks_link.ulnk_url,
+          lnk_type: link.lnks_link.lnk_type
         }
       };
     }