1. not doing work is faster than doing work
2. prefixing attributes was likely causing V8 to not optimize memory layout and execution of annotation queries and fromRaw— by not changing the object shape of the annotation, all annotation attributes get optimized using the same C struct that V8 internally creates

Here's the metrics from the renderer-commonmark performance benchmark with this branch:

The metrics below are taken from the CommonMark specification tests. These are _not_ realistic examples of what you'd find out in the world, but it is a fairly large dataset that runs the code through its paces.

This benchmark was taken on Linux 4.15 on x64 with 2 cores of Intel(R) Xeon(R) CPU.

| Function | Mean | Median | 95th Percentile | Maximum | Standard Deviation |
|----------|------|--------|-----------------|---------|--------------------|
| CommonMarkSource.fromRaw | 0.428ms | 0.367ms | 0.545ms | 14.818923ms | 0.319ms |
| CommonMarkSource.convertTo(OffsetSource) | 0.243ms | 0.196ms | 0.458ms | 22.444555ms | 0.269ms |
| CommonMarkRenderer.render | 0.100ms | 0.084ms | 0.187ms | 4.10044ms | 0.101ms |
| Round trip | 0.817ms | 0.698ms | 1.465ms | 24.706735ms | 0.476ms |

## 🔥 Slow real-life examples

The metrics below are taken from real articles written by Condé Nast editors / writers.
This benchmark was taken on Linux 4.15 on x64 with 2 cores of Intel(R) Xeon(R) CPU.

| Function | Mean | Median | 95th Percentile | Maximum | Standard Deviation |
|----------|------|--------|-----------------|---------|--------------------|
| CommonMarkSource.fromRaw | 12.742ms | 13.337ms | 15.762ms | 25.900293ms | 3.799ms |
| CommonMarkSource.convertTo(OffsetSource) | 38.657ms | 42.174ms | 49.977ms | 50.949202ms | 9.052ms |
| CommonMarkRenderer.render | 36.580ms | 21.329ms | 70.949ms | 78.194161ms | 24.157ms |
| Round trip | 88.133ms | 78.118ms | 133.263ms | 146.801923ms | 34.253ms |

You can compare the results of this to [this file](https://github.com/CondeNast/atjson/blob/latest/packages/%40atjson/renderer-commonmark/performance/README.md)