@tim-evans 2020-01-09
I'm going to close this, since the performance benefits do not outweigh the API change at the moment. We can discuss further in a working group meeting.